requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.23.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
MarkupSafe==3.0.2
pyodbc==5.2.0
requests==2.32.3
tqdm==4.67.1
urllib3==2.4.0
Werkzeug==3.1.3
